refactor message envelopes, support reg. messages
`SendEnvelope` renamed to `DispatchEnvelope`, and its subtypes now
extend to `Registration` types supporting registration and
deregistration of `ActorPath`s.

The new registration types trickle down to the dispatcher who is
responsible for storing these actor paths for lookup, either by their
`UniquePath` or `NamedPath`.
